+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/event/player/ClickEvents.kt
@@ -4,7 +4,9 @@ import dev.aaronhowser.mods.geneticsresequenced.GeneticsResequenced
 import dev.aaronhowser.mods.geneticsresequenced.entity.SupportSlime
 import dev.aaronhowser.mods.geneticsresequenced.gene.behavior.AttributeGenes
 import dev.aaronhowser.mods.geneticsresequenced.gene.behavior.ClickGenes
-import net.minecraft.world.entity.projectile.Projectile
+import dev.aaronhowser.mods.geneticsresequenced.registry.ModItems
+import net.minecraft.world.InteractionHand
+import net.minecraft.world.entity.npc.AbstractVillager
 import net.neoforged.bus.api.SubscribeEvent
 import net.neoforged.fml.common.EventBusSubscriber
 import net.neoforged.neoforge.event.entity.EntityJoinLevelEvent
@@ -22,6 +24,25 @@ object ClickEvents {
         ClickGenes.handleWooly(event)
         ClickGenes.handleMilky(event)
         ClickGenes.handleMeaty(event)
+
+        checkShouldCancel(event)
+    }
+
+    private fun checkShouldCancel(event: PlayerInteractEvent.EntityInteract) {
+        val entity = event.target
+        if (entity !is AbstractVillager) return
+
+        val mainHandStack = event.entity.getItemInHand(InteractionHand.MAIN_HAND)
+        val offHandStack = event.entity.getItemInHand(InteractionHand.OFF_HAND)
+
+        if (
+            mainHandStack.item == ModItems.METAL_SYRINGE.get()
+            || offHandStack.item == ModItems.METAL_SYRINGE.get()
+            || mainHandStack.item == ModItems.SCRAPER.get()
+            || offHandStack.item == ModItems.SCRAPER.get()
+        ) {
+            event.isCanceled = true
+        }
     }
 
     @SubscribeEvent